requirements.txt
streamlit
requests
duckduckgo-search
google-generativeai
python-dotenv
